requirements.txt
# Enhanced Conductor Requirements (Workflow & Documentation System)
python-docx>=1.1.2
markdown>=3.4.0
jinja2>=3.1.3
click>=8.1.7
rich>=13.7.0
requests>=2.31.0
pyyaml>=6.0.1
atlassian-python-api>=3.41.10
PyGithub>=2.1.1
beautifulsoup4>=4.12.2
pydantic-settings>=2.0.0
pydantic>=2.5.3
aiohttp>=3.9.0
